refactor DTO names and location for consistency
