+++ src/main/java/com/chatop/rental/dto/RentalDto.java
@@ -1,9 +1,11 @@
-package com.chatop.rental.dto.responses;
+package com.chatop.rental.dto;
 
 import java.util.Date;
-import java.util.List;
 
+import com.chatop.rental.serialization.CustomDoubleSerializer;
 import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -12,17 +14,22 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class RentalDetailDto {
+public class RentalDto {
     private Integer id;
     private String name;
-    private Integer surface;
-    private Integer price;
-    private List<String> picture;
+    @JsonSerialize(using = CustomDoubleSerializer.class)
+    private Double surface;
+    @JsonSerialize(using = CustomDoubleSerializer.class)
+    private Double price;    
+    private String picture;
     private String description;
     
+    @JsonProperty("owner_id")
     private Integer ownerId;
+    @JsonProperty("created_at")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy/MM/dd", timezone = "UTC")
     private Date createdAt;
+    @JsonProperty("updated_at")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy/MM/dd", timezone = "UTC")
     private Date updatedAt;
 }
