+++ src/main/java/com/chatop/rental/dto/responses/RentalListResponse.java
@@ -2,6 +2,8 @@ package com.chatop.rental.dto.responses;
 
 import java.util.List;
 
+import com.chatop.rental.dto.RentalDto;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
