+++ src/main/java/com/chatop/rental/service/RentalService.java
@@ -3,15 +3,15 @@ package com.chatop.rental.service;
 import java.util.List;
 import java.util.Optional;
 
-import com.chatop.rental.dto.requests.RentalRequest;
+import com.chatop.rental.dto.RentalDto;
+import com.chatop.rental.dto.requests.CreateRentalRequest;
 import com.chatop.rental.dto.requests.UpdateRentalRequest;
 import com.chatop.rental.dto.responses.MessageResponse;
-import com.chatop.rental.dto.responses.RentalDetailDto;
-import com.chatop.rental.dto.responses.RentalDto;
+import com.chatop.rental.dto.responses.RentalDetailResponse;
 
 public interface RentalService {
-	Optional<MessageResponse> createRental(RentalRequest rentalRequest, String userEmail);
+	Optional<MessageResponse> createRental(CreateRentalRequest rentalRequest, String userEmail);
 	List<RentalDto> getAllRentals();
-    Optional<RentalDetailDto> getRentalById(Integer rentalId);
+    Optional<RentalDetailResponse> getRentalById(Integer rentalId);
     Optional<MessageResponse> updateRental(Integer rentalId, UpdateRentalRequest rentalRequest);
 }
